+++ src/main/kotlin/me/ferrandis/textAnalyzer/controller/LanguageController.kt
@@ -1,17 +1,17 @@
 package me.ferrandis.textAnalyzer.controller
 
+import me.ferrandis.textAnalyzer.model.LanguageData
 import me.ferrandis.textAnalyzer.service.LanguageService
-import opennlp.tools.langdetect.Language
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
-import reactor.core.publisher.Flux
+import reactor.core.publisher.Mono
 
 @RestController
 class LanguageController(var service: LanguageService) {
 
     @PostMapping("v1/language")
-    fun dataAnalysis(@RequestBody request: String): Flux<List<Language>> {
+    fun dataAnalysis(@RequestBody request: String): Mono<LanguageData> {
         return service.getLanguage(request);
     }
 }
