+++ src/main/kotlin/me/ferrandis/textAnalyzer/model/LanguageData.kt
@@ -1,0 +1,6 @@
+package me.ferrandis.textAnalyzer.model
+
+import opennlp.tools.langdetect.Language
+
+@JvmRecord
+data class LanguageData(val languages: List<Language>, val text: String)
