+++ src/main/kotlin/me/ferrandis/textAnalyzer/repository/LanguageRepository.kt
@@ -1,0 +1,7 @@
+package me.ferrandis.textAnalyzer.repository
+
+import me.ferrandis.textAnalyzer.model.LanguageData
+import org.springframework.data.repository.reactive.ReactiveCrudRepository
+
+interface LanguageRepository : ReactiveCrudRepository<LanguageData, Long> {
+}
