+++ src/main/kotlin/me/ferrandis/textAnalyzer/repository/SentimentRepository.kt
@@ -1,0 +1,7 @@
+package me.ferrandis.textAnalyzer.repository
+
+import me.ferrandis.textAnalyzer.model.SentimentAnalysis
+import org.springframework.data.repository.reactive.ReactiveCrudRepository
+
+interface SentimentRepository : ReactiveCrudRepository<SentimentAnalysis, Long> {
+}
