+++ src/main/kotlin/me/ferrandis/textAnalyzer/service/AnalysisService.kt
@@ -7,6 +7,7 @@ import edu.stanford.nlp.sentiment.SentimentCoreAnnotations
 import edu.stanford.nlp.sentiment.SentimentCoreAnnotations.SentimentAnnotatedTree
 import edu.stanford.nlp.util.CoreMap
 import me.ferrandis.textAnalyzer.model.SentimentAnalysis
+import me.ferrandis.textAnalyzer.repository.SentimentRepository
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Flux
 import java.util.*
@@ -14,8 +15,8 @@ import java.util.stream.Collectors
 
 
 @Service
-class AnalysisService {
-    fun analyze(request: String) = Flux.just(analyzeAndReturn(request));
+class AnalysisService(var repository: SentimentRepository) {
+    fun analyze(request: String) = analyzeAndReturn(request);
 
     private fun convert(sentence: CoreMap): SentimentAnalysis {
         val tree = sentence.get(SentimentAnnotatedTree::class.java)
@@ -25,14 +26,14 @@ class AnalysisService {
                 sentence.toString())
     }
 
-    private fun analyzeAndReturn(content: String): List<SentimentAnalysis> {
+    private fun analyzeAndReturn(content: String): Flux<SentimentAnalysis> {
         val props = Properties()
         props.setProperty("annotators", "tokenize, ssplit, parse, sentiment")
         val pipeline = StanfordCoreNLP(props)
         val annotation = pipeline.process(content)
-        return annotation.get<List<CoreMap>>(SentencesAnnotation::class.java).stream()
+        return repository.saveAll(annotation.get<List<CoreMap>>(SentencesAnnotation::class.java).stream()
                 .map(this::convert)
-                .collect(Collectors.toList())
+                .collect(Collectors.toList()));
     }
 
 }
