+++ src/main/resources/application.properties
@@ -1,1 +1,3 @@
-
+spring.data.mongodb.host=localhost
+spring.data.mongodb.port=27017
+spring.data.mongodb.database=textAnalyzer
